separate plate tests to new module
